requirements.txt
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
python-dotenv==0.20.0
typing_extensions==4.3.0
Werkzeug==2.2.0
zipp==3.8.1
